cpp: leap exercise update argument types
Comment from @siebenschlaefer on the usage of unsigned for year.

[ES.100-102 of the C++ Core
Guidelines](http://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#es100-dont-mix-signed-and-unsigned-arithmetic)
talk about signed and unsigned types. They basically state that you
should use signed types for arithmetic, unsigned types for
bit manipulation, and that you should never mix them.

There's a video of the Interactive Panel: Ask Us Anything from the
"Going Native 2013" conference where Chandler Carruth, Bjarne
Stroustrup, and Herb Sutter talked about signed and unsigned integers
(timestamps 12:13 and 42:40). Later they even jokingly apologized for
the use of unsigned integers in the standard library (timestamp
1:02:50).

Even if you use an unsigned type for the argument the user could still
pass a negative integer to the function and that's a perfectly valid
thing to do because the conversion from signed to unsigned is well
defined. But the result of the function would probably be unexpected.